Load National Reference Station chemistry data

Load water chemistry measurements from NRS stations, including nutrients (nitrogen, phosphorus, silicate), dissolved oxygen, carbon chemistry, and physical properties. Data are quality controlled using IMOS flags.

pr_get_NRSChemistry()

Value

A dataframe in long format with chemistry data including:

StationCode, StationName - Station identifiers

SampleTime_Local - Local sampling date and time

SampleDepth_m - Sampling depth

Parameters - Chemistry parameter name

Values - Measured or calculated value

Details

The function retrieves and processes the following parameters:

Nutrients:

Nitrate_umolL - Nitrate concentration (µmol/L)

Nitrite_umolL - Nitrite concentration (µmol/L)

Ammonium_umolL - Ammonium concentration (µmol/L)

Phosphate_umolL - Phosphate concentration (µmol/L)

Silicate_umolL - Silicate concentration (µmol/L)

NOx_umolL - Nitrogen oxides (nitrate + nitrite, µmol/L)

DIN_umolL - Dissolved inorganic nitrogen (NOx + ammonium, µmol/L)

Carbon Chemistry:

DIC_umolkg - Dissolved inorganic carbon (µmol/kg)

Alkalinity_umolkg - Total alkalinity (µmol/kg)

Other:

Oxygen_umolL - Dissolved oxygen (µmol/L)

Salinity - Salinity (unitless, PSU)

SecchiDepth_m - Secchi depth, measure of water clarity (m)

Redfield - N:P ratio (calculated as NOx/Phosphate)

Quality control flags are applied using pr_apply_Flags(). Samples are filtered to include only NRS stations.

NOx, DIN, and Redfield ratio are calculated from the measured nutrient values.

See also

Examples

df <- pr_get_NRSChemistry()

# Examine available parameters
unique(df$Parameters)
#>  [1] "SecchiDepth_m"     "Silicate_umolL"    "Phosphate_umolL"  
#>  [4] "Ammonium_umolL"    "Nitrate_umolL"     "Nitrite_umolL"    
#>  [7] "Oxygen_umolL"      "DIC_umolkg"        "Alkalinity_umolkg"
#> [10] "Salinity"          "NOx_umolL"         "DIN_umolL"        
#> [13] "Redfield"         

# Filter for surface nitrate at specific stations
df_nitrate <- df %>%
  dplyr::filter(Parameters == "Nitrate_umolL", 
                SampleDepth_m == 0,
                StationCode %in% c("MAI", "PHB"))