Import NRS sampling trip metadata

Load metadata for all National Reference Station sampling trips, including dates, locations, and environmental conditions at the time of sampling.

pr_get_NRSTrips()

Value

A dataframe with NRS trip information including:

TripCode - Unique trip identifier

StationCode, StationName - Station information

SampleTime_Local - Local sampling date and time

Year_Local, Month_Local - Temporal components

Latitude, Longitude - Geographic coordinates

Additional metadata fields

Details

Returns information about each NRS sampling trip (voyage), including the trip code, station, date/time, and basic metadata. This is useful for understanding sampling effort over time and for joining with other datasets.

Only NRS project trips are included (SOTS trips are excluded). Samples designated as 'P' (plankton) samples or with no sample type designation are included.

See also

Examples

df <- pr_get_NRSTrips()

# Examine sampling frequency by station
table(df$StationName, df$Year_Local)
#>                          
#>                           2002 2003 2004 2005 2006 2007 2008 2009 2010 2011
#>   Bonney Coast               0    0    0    0    0    0    0    0    0    0
#>   Darwin                     0    0    0    0    0    0    0    0    0    2
#>   Esperance                  0    0    0    0    0    0    0    4    3    4
#>   Kangaroo Island            0    0    0    0    0    0    8    9    9   11
#>   Maria Island               0    0    0    0    0    0    0    9    9   10
#>   Ningaloo                   0    0    0    0    0    0    0    0    1    4
#>   North Stradbroke Island    0    0    0    0    0    0    4   15   12   12
#>   Port Hacking               0    0    0    0    0    0    0   13   13   12
#>   Port Hacking 4            11   12   12    9    5   11   11    1    0    0
#>   Rottnest Island            0    0    0    0    0    0    0    2    9   11
#>   Yongala                    0    0    0    0    0    0    0    3    9   11
#>                          
#>                           2012 2013 2014 2015 2016 2017 2018 2019 2020 2021
#>   Bonney Coast               0    0    0    0    0    0    0    0    0    0
#>   Darwin                    14    6   11   11   10   10   10   10   11   10
#>   Esperance                  5    2    0    0    0    0    0    0    0    0
#>   Kangaroo Island            3    4    7    6    4    4    4    4    3    4
#>   Maria Island              11   11   13    9   10   11   10   10    9   12
#>   Ningaloo                   4    3    0    0    0    0    0    0    0    0
#>   North Stradbroke Island    9   12   11   12   10   11   10   10   10   11
#>   Port Hacking              10   12   12   10   10   10   10   11    9    9
#>   Port Hacking 4             0    0    0    0    0    0    0    0    0    0
#>   Rottnest Island           11   10   11   10   11   12   12   12   11   11
#>   Yongala                   12   11   12   12   12   12   12   12   12   12
#>                          
#>                           2022 2023 2024 2025
#>   Bonney Coast               0    0    3    3
#>   Darwin                    10   10   10    6
#>   Esperance                  0    0    0    0
#>   Kangaroo Island            4    3    4    3
#>   Maria Island              11   10   11   10
#>   Ningaloo                   0    0    0    0
#>   North Stradbroke Island   10   12    8    5
#>   Port Hacking              11    9   10    7
#>   Port Hacking 4             0    0    0    0
#>   Rottnest Island           10   11   10    9
#>   Yongala                   11   13   12    8