Combined timeseries and climatology plots for environmental variables

Combined timeseries and climatology plots for environmental variables

pr_plot_env_var(df, pal = "matter", trend = "None", Scale = "identity")

Arguments

df

A dataframe from pr_get_nuts or pr_get_pigs

pal

A Palette from cmocean

trend

Trend line to be used, options None, Smoother, Linear

Scale

scale on which to plot y axis

Value

A plotly plot with timeseries and climatology at depths

Examples

df <- pr_get_nuts() %>% pr_plot_env_var()
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: the condition has length > 1 and only the first element will be used
#> Warning: Removed 1337 rows containing non-finite values (stat_smooth).
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  5.0933e-17
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  5.0933e-17
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 4.97
#> Warning: neighborhood radius 6.03
#> Warning: reciprocal condition number  3.1723e-17
#> Warning: There are other near singularities as well. 36.361
#> Warning: pseudoinverse used at 4.97
#> Warning: neighborhood radius 6.03
#> Warning: reciprocal condition number  3.1723e-17
#> Warning: There are other near singularities as well. 36.361
#> Warning: pseudoinverse used at 1.97
#> Warning: neighborhood radius 5.03
#> Warning: reciprocal condition number  1.6108e-16
#> Warning: There are other near singularities as well. 16.241
#> Warning: pseudoinverse used at 1.97
#> Warning: neighborhood radius 5.03
#> Warning: reciprocal condition number  1.6108e-16
#> Warning: There are other near singularities as well. 16.241
#> Warning: pseudoinverse used at 0.97
#> Warning: neighborhood radius 6.03
#> Warning: reciprocal condition number  1.0408e-16
#> Warning: There are other near singularities as well. 36.361
#> Warning: pseudoinverse used at 0.97
#> Warning: neighborhood radius 6.03
#> Warning: reciprocal condition number  1.0408e-16
#> Warning: There are other near singularities as well. 36.361
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 0.97
#> Warning: neighborhood radius 6.03
#> Warning: reciprocal condition number  6.6579e-17
#> Warning: There are other near singularities as well. 36.361
#> Warning: pseudoinverse used at 0.97
#> Warning: neighborhood radius 6.03
#> Warning: reciprocal condition number  6.6579e-17
#> Warning: There are other near singularities as well. 36.361
#> Warning: pseudoinverse used at 0.975
#> Warning: neighborhood radius 5.025
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 25.251
#> Warning: pseudoinverse used at 0.975
#> Warning: neighborhood radius 5.025
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 25.251
#> Warning: pseudoinverse used at 3.985
#> Warning: neighborhood radius 3.015
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 9.0902
#> Warning: pseudoinverse used at 3.985
#> Warning: neighborhood radius 3.015
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 9.0902
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  5.0933e-17
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  5.0933e-17
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 2.995
#> Warning: neighborhood radius 1.005
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 1.01
#> Warning: pseudoinverse used at 2.995
#> Warning: neighborhood radius 1.005
#> Warning: reciprocal condition number  0
#> Warning: There are other near singularities as well. 1.01
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9
#> Warning: pseudoinverse used at 5
#> Warning: neighborhood radius 3
#> Warning: reciprocal condition number  1.3233e-16
#> Warning: There are other near singularities as well. 9